requirements.txt
Flask
gunicorn
google-cloud-bigquery